Revert "Merge pull request #3 from carlivas/timestep"
This reverts commit 2e22eacb7e87bbc5db681d1e75e578a2ff1d58f0, reversing
changes made to 3cbfd3b20c38488db7868379d1855fdcf5a8f279.

+++ scripts/python/mods/simulation.py
@@ -178,10 +178,11 @@ class Simulation:
         prev_mod_state = prev_t % self.state_buffer.skip
         mod_state = self.t % self.state_buffer.skip
         do_save_state = prev_mod_state > mod_state
-
         prev_mod_density_field = prev_t % self.density_field_buffer.skip
         mod_density_field = self.t % self.density_field_buffer.skip
         do_save_density_field = prev_mod_density_field > mod_density_field
+        # do_save_density_field = np.isclose(self.t % self.density_field_buffer.skip, 0) or any(
+        #     np.isclose(self.t, self.density_field_buffer.preset_times))
 
         if do_save_state:
             self.state_buffer.add(state=self.get_state(), t=self.t)
@@ -548,7 +549,7 @@ class Simulation:
         ax.set_ylim(-self.half_height, self.half_height)
         ax.set_aspect('equal', 'box')
         if t is not None:
-            t = float(round(t, 2))
+            t = round(t, 2)
             ax.set_title(f'{t=}', fontsize=7)
         else:
             ax.set_title('')
